+++ NEWS.md
@@ -3,6 +3,7 @@
 - Fix `api_version` in `create_ssh_key()`
 - rename `ssh_key*` functions to `checkout_key*`
 - `create_checkout_key()` change default for arg `type` from "github-user-key" to "deploy-key"
+- add argument `encode` to `circleHTTP()`
 
 # circleci 0.1.0